% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{pruneRegulon}
\alias{pruneRegulon}
\title{Prune Regulons}
\usage{
pruneRegulon(regulon, cutoff = 50, eliminate = FALSE)
}
\arguments{
\item{regulon}{Object of class regulon}

\item{cutoff}{Number indicating the maximum size for the regulons (maximum number of target genes)}

\item{eliminate}{Logical whether regulons smalles than \code{cutoff} should be eliminated}
}
\value{
Prunned regulon
}
\description{
This function limits the maximum size of the regulons
}
\examples{
data(bcellViper, package="bcellViper")
hist(sapply(regulon, function(x) sum(x$likelihood)/max(x$likelihood)), nclass=20)
preg <- pruneRegulon(regulon, 400)
hist(sapply(preg, function(x) sum(x$likelihood)/max(x$likelihood)), nclass=20)
}
\seealso{
\code{\link{viper}}, \code{\link{msviper}}
}

